compare readers
