+++ chapter-06/inprogress/go.mod
@@ -1,3 +1,5 @@
 module github.com/ablqk/tiny-go-projects/chapter-06/inprogress
 
 go 1.19
+
+require golang.org/x/exp v0.0.0-20230118134722-a68e582fa157
